requirements.txt
Django==1.6.5
MySQL-python==1.2.5
django-extensions==1.3.10
django-adminplus==0.3
django-mama-cas==1.0.1
South==1.0.2
django-supervisor==0.3.2
django-simple-captcha==0.4.2
